added description for updating `.condarc` file
a conda environment created at a not-default location does not show up in the list of environments obtained with `conda env list`. For such environments we need to activate them differently so as to not see their entire path in place of the `(env_name)` which identifies which environment is currently active. Added documentation on what must be done in such cases.